requirements.txt
peewee
pyyaml
dependency-injector
PyMySQL
mongoengine
